refactor: restructure external libs and deps
