+++ apps/client/.svelte-kit/types/src/routes/$types.d.ts
@@ -10,7 +10,7 @@ type EnsureDefined<T> = T extends null | undefined ? {} : T;
 type OptionalUnion<U extends Record<string, any>, A extends keyof U = U extends U ? keyof U : never> = U extends unknown ? { [P in Exclude<A, keyof U>]?: never } & U : never;
 export type Snapshot<T = any> = Kit.Snapshot<T>;
 type PageParentData = EnsureDefined<LayoutData>;
-type LayoutRouteId = RouteId | "/" | "/about" | "/about/[slug=slug]" | "/auth" | "/auth/new-user" | "/authored" | "/authored/edit" | "/chat" | "/courses" | "/courses/[slug]" | "/courses/[slug]/[lesson]" | "/explore" | "/issues" | "/notifications" | "/profile" | "/settings" | null
+type LayoutRouteId = RouteId | "/" | "/about" | "/auth" | "/auth/new-user" | "/authored" | "/authored/edit" | "/chat" | "/courses" | "/courses/[slug]" | "/courses/[slug]/[lesson]" | "/explore" | "/issues" | "/notifications" | "/profile" | "/settings" | null
 type LayoutParams = RouteParams & { slug?: string,lesson?: string }
 type LayoutServerParentData = EnsureDefined<{}>;
 type LayoutParentData = EnsureDefined<{}>;
