+++ apps/client/.svelte-kit/types/src/routes/proxy+layout.server.ts
@@ -1,19 +1,7 @@
 // @ts-nocheck
 import type { LayoutServerLoad } from './$types';
-import apolloClient, { githubClient } from "$lib/apollo";
-import {
-  PUBLIC_HASURA_ADMIN_SECRET,
-  PUBLIC_GITHUB_ACCESS_CODE,
-  PUBLIC_GITHUB_API_ENDPOINT,
-} from "$env/static/public";
 
-type ServerData = {
-  session?: any,
-  githubClient?: any,
-}
-
-
-export const load = async (event: Parameters<LayoutServerLoad>[0])=> {
+export const load = async (event: Parameters<LayoutServerLoad>[0]) => {
   const session = await event.locals.getSession();
 
   return {
