+++ apps/client/src/lib/graphql/mutations/courses.ts
