+++ apps/client/src/lib/graphql/queries/courses.ts
