+++ apps/client/src/lib/graphql/queries/lessons.ts
